% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-coefs.R
\name{add_colours}
\alias{add_colours}
\title{Function for adding colours and species traits}
\usage{
add_colours(
  coefs,
  col_var = "group",
  species_data = stats,
  add_spp_data = TRUE,
  manual_colours = FALSE,
  last_used = FALSE
)
}
\arguments{
\item{coefs}{Coefficient dataframe from TMB model output.}

\item{col_var}{Variable to colour by.}

\item{species_data}{Species trait dataframe.}

\item{add_spp_data}{Logical for if above is to be joined.}

\item{manual_colours}{Logical for if custom colours in this function.}

\item{last_used}{Apply colours from previous model.}
}
\description{
Function for adding colours and species traits
}
