% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-data-processing.R
\name{add_missing_depths}
\alias{add_missing_depths}
\title{Add missing depths}
\usage{
add_missing_depths(dat, survey, years, bath)
}
\arguments{
\item{bath}{Output from [interp_survey_bathymetry()].}
}
\description{
Add missing depths
}
