% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcslope.R
\name{calcslope}
\alias{calcslope}
\title{Calculate trend in varible for pixels in a raster brick}
\usage{
calcslope(rx, delta_t_step = 2, na.rm = TRUE)
}
\arguments{
\item{rx}{Raster brick containing layers for multiple years.}

\item{na.rm}{Logical for if NAs are removed.}

\item{delta_t}{Number of time steps between layers, 
if not 1 allows slope to be calculated for a single unit of time.}
}
\description{
Calculate trend in varible for pixels in a raster brick
}
