% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{chopstick_slopes}
\alias{chopstick_slopes}
\title{Save slopes from chopstick plots}
\usage{
chopstick_slopes(
  model,
  type = NULL,
  interaction_column = "temp_trend_scaled:mean_temp_scaled",
  x_variable = "temp_trend_scaled",
  species = NULL
)
}
\description{
Save slopes from chopstick plots
}
