% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-coefs.R
\name{coef_scatterplot}
\alias{coef_scatterplot}
\title{Plot coefficients against species traits}
\usage{
coef_scatterplot(
  model_coefs,
  x,
  coef = c("temp_trend_scaled", "DO_trend_scaled"),
  group = "age",
  point_size = 0.75,
  point_alpha = 0.65,
  point_shapes = c(21, 19),
  pointrange = T,
  regression = T
)
}
\arguments{
\item{model_coefs}{Coefficient dataframe joined with species traits.}

\item{x}{Trait variable for x-axis.}

\item{coef}{Coefficient to plot.}

\item{group}{Colour by grouping.}

\item{point_size}{}

\item{point_alpha}{}

\item{point_shapes}{Vector of shapes. Defaults to open and closed circles.}

\item{regression}{Logical for adding regression line.}
}
\description{
Plot coefficients against species traits
}
