% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc-gradient.R
\name{collapse_outliers}
\alias{collapse_outliers}
\title{Collapse outliers produced by velocity calculations}
\usage{
collapse_outliers(.x, outliers)
}
\arguments{
\item{.x}{Vector to be collapsed}

\item{outliers}{Vector of quantile thresholds used in collapse}
}
\description{
Collapse outliers produced by velocity calculations
}
