% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc.R
\name{dist_based_vocc}
\alias{dist_based_vocc}
\title{Velocity of climate change based on distance}
\usage{
dist_based_vocc(
  start_data,
  end_data,
  x = "x",
  y = "y",
  variable_names = c("var.1", "var.2"),
  round_fact = NULL,
  min_thresholds = NULL,
  max_thresholds = NULL,
  plus_minus = c(0.5, 0.5),
  match_logic = NULL,
  cell_size = 2,
  max_dist = NULL,
  min_dist = cell_size/2,
  delta_t = 10,
  raster = FALSE,
  kNN = FALSE
)
}
\arguments{
\item{start_data}{Named list containing starting climate data.}

\item{end_data}{Named list containing target climate data.}

\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}

\item{variable_names}{Vector of column/layer names within each data element.}

\item{round_fact}{Speed up searches by rounding (1 = integers; 10 = 10ths; 100 = 100ths).
If NULL, will use 10x precision of the plus_minus threshold when `match_logic` = NULL.}

\item{min_thresholds}{Optional vector of negative thresholds.
Include if sensitivity to climate is not symmetrical and `match_logic` = NULL.}

\item{max_thresholds}{Optional vector of positive thresholds.
Include if sensitivity to climate is not symmetrical and `match_logic` = NULL.}

\item{plus_minus}{Vector of plus/minus threshold(s) to define a symmetical (default = 1 unit) climate match.}

\item{match_logic}{An optional vector of logical functions applied using 'rounding' of allnclimate values.
If max_thresholds are not provided, the default of 'NULL' applies symmetrical plus/minus thresholds.}

\item{cell_size}{Cell size in raster or distance to nearest data point.}

\item{max_dist}{Distance reported when no analogue found. Default will
estimate this assuming that both major axes are likely to hold an analogue.}

\item{min_dist}{Defaults to half cell size.}

\item{delta_t}{Time difference between starting and target time periods.}

\item{raster}{Logical for whether climate data is in raster form.}

\item{kNN}{Logical for whether to use kNN search for extra speed
(but uses rounding, symmetrical thresholds, and only returns 1 target cell per source).}
}
\description{
Velocity of climate change based on distance
}
