% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-conditional-effects.R
\name{effect_plot}
\alias{effect_plot}
\title{Title}
\usage{
effect_plot(
  model,
  pred,
  pred.values = NULL,
  centered = "all",
  plot.points = FALSE,
  interval = FALSE,
  data = NULL,
  at = NULL,
  int.type = c("confidence", "prediction"),
  int.width = 0.95,
  outcome.scale = "response",
  robust = FALSE,
  cluster = NULL,
  vcov = NULL,
  set.offset = 1,
  x.label = NULL,
  y.label = NULL,
  pred.labels = NULL,
  main.title = NULL,
  colors = "black",
  line.thickness = 1.1,
  point.size = 1.5,
  point.alpha = 0.6,
  point.color = "black",
  point.color.var = NULL,
  jitter = 0,
  rug = FALSE,
  rug.sides = "lb",
  force.cat = FALSE,
  cat.geom = c("point", "line", "bar"),
  cat.interval.geom = c("errorbar", "linerange"),
  cat.pred.point.size = 3.5,
  partial.residuals = FALSE,
  color.class = colors,
  ...
)
}
\arguments{
\item{model}{}

\item{pred}{}

\item{pred.values}{}

\item{centered}{}

\item{plot.points}{}

\item{interval}{}

\item{data}{}

\item{at}{}

\item{int.type}{}

\item{int.width}{}

\item{outcome.scale}{}

\item{robust}{}

\item{cluster}{}

\item{vcov}{}

\item{set.offset}{}

\item{x.label}{}

\item{y.label}{}

\item{pred.labels}{}

\item{main.title}{}

\item{colors}{}

\item{line.thickness}{}

\item{point.size}{}

\item{point.alpha}{}

\item{point.color}{}

\item{jitter}{}

\item{rug}{}

\item{rug.sides}{}

\item{force.cat}{}

\item{cat.geom}{}

\item{cat.interval.geom}{}

\item{cat.pred.point.size}{}

\item{partial.residuals}{}

\item{color.class}{}

\item{...}{}
}
\value{

}
\description{
Title
}
