% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{fixed_density}
\alias{fixed_density}
\title{Calculate density curve for non-time-varying, scaled, quadratic parameters}
\usage{
fixed_density(
  m,
  predictor = "temp",
  time = "year",
  fixed_params = 1,
  quadratics_only = TRUE
)
}
\arguments{
\item{m}{Output of sdmTMB model}

\item{predictor}{Prefix for scaled parameter in model}

\item{time}{Variable containing time_steps with separate intercepts}

\item{fixed_params}{}

\item{quadratics_only}{If True doesn't plot 'happy face' quadratics.}
}
\description{
Calculate density curve for non-time-varying, scaled, quadratic parameters
}
