% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{get_optimal_value}
\alias{get_optimal_value}
\title{Find optimal value from density curve}
\usage{
get_optimal_value(dat, xlimits = c(0, max(dat$x)))
}
\arguments{
\item{dat}{Dataframe resulting from any of the above density functions}

\item{xlimits}{Set prior for possible values}
}
\description{
Find optimal value from density curve
}
