% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{get_quadratic_roots}
\alias{get_quadratic_roots}
\title{Find roots of quadratic fixed effect curves}
\usage{
get_quadratic_roots(
  m,
  predictor = "do_mlpl",
  fixed_param = 1,
  threshold,
  plot = FALSE,
  quadratic_only = TRUE
)
}
\arguments{
\item{m}{Output of sdmTMB model}

\item{predictor}{Prefix for scaled parameter in model}

\item{fixed_param}{Number in sequence if model contains multiple quadratic fixed effects}

\item{threshold}{What proportion of total y to calculate roots for}

\item{plot}{Logical for plotting quadratic}
}
\description{
Find roots of quadratic fixed effect curves
}
