% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-data-processing.R
\name{interp_survey_bathymetry}
\alias{interp_survey_bathymetry}
\title{Interpolate survey bathymetry}
\usage{
interp_survey_bathymetry(dat, utm_zone = 9)
}
\description{
Interpolate survey bathymetry
}
