% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc-gradient.R
\name{make_gradient_brick}
\alias{make_gradient_brick}
\title{Create a RasterBrick from gridded predictions}
\usage{
make_gradient_brick(data, layer, scale_fac = 1, time_step = "year")
}
\description{
Create a RasterBrick from gridded predictions
}
