% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal-data.R
\name{make_raster_brick}
\alias{make_raster_brick}
\title{Make raster brick from dataframe of predicted values through time}
\usage{
make_raster_brick(data, parameter = "est", scale_fac = 1, time_var = "year")
}
\arguments{
\item{data}{Dataframe of spatial grid for multiple time slices.}

\item{parameter}{Name of column with raster values.}

\item{scale_fac}{Controls how the original projection is aggregated.}

\item{time_var}{Name of column with time variable.}
}
\description{
Make raster brick from dataframe of predicted values through time
}
