% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-trend-data.R
\name{make_trend_data}
\alias{make_trend_data}
\title{Make trend data from sdmTMB output predict funciton}
\usage{
make_trend_data(
  data,
  ssid = NULL,
  start_time = NULL,
  end_time = NULL,
  skip_time = NULL,
  input_cell_size = 2,
  scale_fac = 1,
  time_var = "year",
  delta_t_total = 10,
  delta_t_step = 2,
  indices = c(1, 2),
  variable_names = "est"
)
}
\arguments{
\item{data}{List of dataframes each containing output from predict.}

\item{ssid}{Vector of ssid to be included.}

\item{start_time}{Starting time for VOCC calculation. Default is NULL and all data is used.}

\item{end_time}{End time for VOCC calculation. Default is NULL and all data is used.}

\item{skip_time}{Time steps to be excluded. Default is NULL and all data is used.}

\item{input_cell_size}{Cell size of predict grid.}

\item{scale_fac}{Factor by which cells are aggregated for VOCC.}

\item{time_var}{Name of column containing time data.}

\item{delta_t_total}{Mean time between start and end values.}

\item{delta_t_step}{Time between each time step in same units as above.}

\item{indices}{Vector of length equal to number of time steps retained for analysis,
where 1 = starting time step(s), 2 = end time step(s).}

\item{variable_names}{Name(s) of column containing parameter(s).}
}
\description{
Make trend data from sdmTMB output predict funciton
}
