% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-vector-data.R
\name{make_vector_data}
\alias{make_vector_data}
\title{Make VOCC vector data from sdmTMB output predict funciton}
\usage{
make_vector_data(
  data,
  ssid = NULL,
  start_time = NULL,
  end_time = NULL,
  skip_time = NULL,
  input_cell_size = 2,
  scale_fac = 1,
  min_dist = input_cell_size * scale_fac/2,
  time_var = "year",
  delta_t_total = 10,
  delta_t_step = 2,
  indices = c(1, 2),
  variable_names = "est",
  round_fact = NULL,
  min_thresholds = NULL,
  max_thresholds = NULL,
  plus_minus = c(0.5),
  match_logic = NULL
)
}
\arguments{
\item{data}{List of dataframes each containing output from predict.}

\item{ssid}{Vector of ssid to be included.}

\item{start_time}{Starting time for VOCC calculation. Default is NULL and all data is used.}

\item{end_time}{End time for VOCC calculation. Default is NULL and all data is used.}

\item{skip_time}{Time steps to be excluded. Default is NULL and all data is used.}

\item{input_cell_size}{Cell size of predict grid.}

\item{scale_fac}{Factor by which cells are aggregated for VOCC.}

\item{time_var}{Name of column containing time data.}

\item{delta_t_total}{Mean time between start and end values.}

\item{delta_t_step}{Time between each time step in same units as above.}

\item{indices}{Vector of length equal to number of time steps retained for analysis,
where 1 = starting time step(s), 2 = end time step(s).}

\item{variable_names}{Name(s) of column containing parameter(s).}

\item{round_fact}{Speed up searches by rounding (1 = integers; 10 = 10ths; 100 = 100ths).
If NULL, will use 10x precision of the plus_minus threshold when `match_logic` = NULL.}

\item{min_thresholds}{Optional vector of negative thresholds. Apply 'Inf' for no min threshold. 
Include if sensitivity to the direction of climate change is not symmetrical and `match_logic` = NULL.}

\item{max_thresholds}{Optional vector of positive thresholds. Apply 'Inf' for no max threshold. 
Include if sensitivity to the direction of climate change is not symmetrical and `match_logic` = NULL.}

\item{plus_minus}{Vector of plus/minus threshold(s) to define a symmetical (default = 1 unit) climate match.}

\item{match_logic}{An optional vector of logical functions applied using 'rounding' of allnclimate values.
If max_thresholds are not provided, the default of 'NULL' applies symmetrical plus/minus thresholds.}
}
\description{
Make VOCC vector data from sdmTMB output predict funciton
}
