% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_chop_est}
\alias{plot_chop_est}
\title{Plot estimates from ends of chopsticks}
\usage{
plot_chop_est(
  model,
  type = NULL,
  x_variable = "temp_trend_scaled",
  where_on_x = "max",
  sort_var = "mean",
  alt_order = F,
  sort_where_on_x = "max",
  legend_position = c(0.7, 0.95),
  point_size = 0.75,
  alpha_range = c(0.5, 0.99),
  add_grey_bars = F,
  name_chop_type = T,
  colours = NULL
)
}
\arguments{
\item{model}{}

\item{type}{Select type if more than one interaction in model}

\item{x_variable}{Which variable on x-axis}

\item{where_on_x}{Where on the x-axis to select estimate to plot. 
Options include max, min, mean.}

\item{sort_var}{Defaults to "mean", current alternatives are "NULL" or "range".}

\item{alt_order}{Logical, if plots will be ordered by mean for alternate x value.}

\item{sort_where_on_x}{Can either sort by mean estimate for "max" x_variable or "min" x_variable.}

\item{legend_position}{Give legend position}

\item{point_size}{Change point size}

\item{alpha_range}{Choose alpha for non-sig and sig slopes}

\item{add_grey_bars}{Add grey stripes to alternating species}

\item{name_chop_type}{Add type to chopstick name in legend}

\item{colours}{Add custom colours.}
}
\description{
Plot estimates from ends of chopsticks
}
