% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-conditional-effects.R
\name{plot_effect_continuous}
\alias{plot_effect_continuous}
\title{Title}
\usage{
plot_effect_continuous(
  predictions,
  pred,
  plot.points = FALSE,
  interval = FALSE,
  data = NULL,
  x.label = NULL,
  y.label = NULL,
  pred.labels = NULL,
  main.title = NULL,
  colors = NULL,
  line.thickness = 1.1,
  jitter = 0.1,
  resp = NULL,
  weights = NULL,
  rug = FALSE,
  rug.sides = "b",
  point.size = 1,
  point.alpha = 0.6,
  point.color = "black",
  point.color.var = point.color.var
)
}
\arguments{
\item{predictions}{}

\item{pred}{}

\item{plot.points}{}

\item{interval}{}

\item{data}{}

\item{x.label}{}

\item{y.label}{}

\item{pred.labels}{}

\item{main.title}{}

\item{colors}{}

\item{line.thickness}{}

\item{jitter}{}

\item{resp}{}

\item{weights}{}

\item{rug}{}

\item{rug.sides}{}

\item{point.size}{}

\item{point.alpha}{}

\item{point.color}{}
}
\value{

}
\description{
Title
}
