% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-maps.R
\name{plot_facet_map}
\alias{plot_facet_map}
\title{Raster map with year facets}
\usage{
plot_facet_map(
  df,
  column = "est",
  X = "X",
  Y = "Y",
  viridis_option = "C",
  white_zero = FALSE,
  low_fill = "Steel Blue 4",
  mid_fill = "white",
  high_fill = "Red 3",
  na_colour = "red",
  transform_col = no_trans,
  raster_limits = NULL,
  text_size = 4,
  legend_position = "right"
)
}
\arguments{
\item{df}{Dataframe.}

\item{column}{Name column to be plotted.}
}
\description{
Raster map with year facets
}
