% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_fuzzy_chopsticks}
\alias{plot_fuzzy_chopsticks}
\title{Plot interactions with confidence intervals}
\usage{
plot_fuzzy_chopsticks(
  model,
  y_label = "Predicted biomass trend",
  x_variable = "temp_trend_scaled",
  type = NULL,
  colours = NULL,
  line_size = 0.7,
  alpha_range = c(0.25, 0.9),
  global_col = "gray30",
  choose_species = NULL,
  choose_age = NULL,
  order_by_chops = c("low", "high"),
  slopes = NULL,
  imm_model = NULL,
  imm_slopes = NULL,
  rug = F
)
}
\arguments{
\item{model}{TMB model with X_pj and pred_dat elements}

\item{y_label}{Label response variable}

\item{x_variable}{Variable to plot on x axis}

\item{type}{Interaction type, if more than one in model}

\item{colours}{Default NULL gives red for high and blue for low}

\item{line_size}{Change line thickness}

\item{alpha_range}{Choose alpha for non-sig and sig slopes}

\item{global_col}{Set colour for global chopsticks}

\item{choose_species}{Option to choose just one species}

\item{choose_age}{Option to choose just one age class}

\item{slopes}{Df of slopes from chopstick_slopes function}

\item{imm_model}{Add immature data from separate model}

\item{imm_slopes}{Add separate slope df for immatures}

\item{rug}{Logical for adding rug}

\item{order_var}{Varible to order plots by}
}
\description{
Plot interactions with confidence intervals
}
\examples{
plot_fuzzy_chopsticks(model,
  y_label = "Predicted biomass trend",
  x_variable = "temp_trend_scaled",
  type = "temp"
)
}
