% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_raw_chopsticks}
\alias{plot_raw_chopsticks}
\title{Plot raw chopsticks for vocc regression models}
\usage{
plot_raw_chopsticks(
  model,
  species = NULL,
  variables = c("mean_temp_scaled", "squashed_temp_vel_scaled"),
  choose_x = NULL
)
}
\arguments{
\item{model}{Model from vocc_regression function.
Requires only named data element and coefs table.}

\item{species}{Species effect to be plotted.}

\item{variables}{List which variables are interacting.}

\item{choose_x}{Which varible from list is on x-axis (either 1 or 2)}
}
\description{
Plot raw chopsticks for vocc regression models
}
