% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcslope.R
\name{raster_slopes}
\alias{raster_slopes}
\title{Calculates slope for regression line through values of raster layers for each cell}
\usage{
raster_slopes(x)
}
\arguments{
\item{x}{Cell of a raster as provided by `raster::calc(rbrick, raster_slopes)`}
}
\description{
Calculates slope for regression line through values of raster layers for each cell
}
