% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{slope_scatterplot}
\alias{slope_scatterplot}
\title{Scatterplot of slopes against traits}
\usage{
slope_scatterplot(
  slopes_w_traits,
  x,
  slope_var = "slope_est",
  col_group = "chopstick",
  shape_group = "age",
  point_size = 0.75,
  point_alpha = 0.99,
  point_shapes = c(21, 19),
  pointrange = T,
  regression = F
)
}
\arguments{
\item{slopes_w_traits}{Df of slopes merged with df of species traits}

\item{x}{Variable (species trait) on x-axis}

\item{slope_var}{Varible containing slope estimate}

\item{col_group}{Varible to colour by}

\item{shape_group}{Varible to set shape by}

\item{point_size}{Set point size}

\item{point_alpha}{Set point alpha}

\item{point_shapes}{Vector of shapes. Defaults to open and closed circles}

\item{pointrange}{Logical for adding 95CI bars to points. Default T}

\item{regression}{Logical for plotting simple linear regression line}
}
\description{
Scatterplot of slopes against traits
}
