% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal-data.R
\name{spatiotemporal_grid}
\alias{spatiotemporal_grid}
\title{Create prediction grid for each year}
\usage{
spatiotemporal_grid(
  data,
  ssid = NULL,
  survey_abbrev = NULL,
  dummy_year,
  cell_width = 2
)
}
\arguments{
\item{data}{Original dataframe of spatiotemporal data.}

\item{ssid}{Survey series id if using established survey grid.}

\item{survey_abbrev}{Survey series abbreviation if using established survey grid.}

\item{dummy_year}{Single year (for each area) on which to base the survey grid.}
}
\description{
Create prediction grid for each year
}
