% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{time_varying_density}
\alias{time_varying_density}
\title{Calculate density curve for time-varying parameters}
\usage{
time_varying_density(m, predictor = "depth")
}
\arguments{
\item{m}{Output of sdmTMB model}

\item{predictor}{Prefix for scaled parameter in model}
}
\description{
Calculate density curve for time-varying parameters
}
