% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-vector-data.R
\name{trim_vector_data}
\alias{trim_vector_data}
\title{Trim VOCC output to only include vectors for cells with specific end conditions}
\usage{
trim_vector_data(
  data,
  variable_names,
  lower_change,
  upper_change,
  lower_thresholds = NULL,
  upper_thresholds = NULL,
  cell_size = 2,
  dist_intercept = cell_size,
  max_dist = max(data$distance, na.rm = TRUE),
  min_dist = cell_size/2
)
}
\arguments{
\item{data}{Dataframe created by make-vector-data function}

\item{variable_names}{Names of climate variable(s) used in make-vector-data function}

\item{lower_change}{When not c(Inf), cells will be trimmed based on lower_thresholds}

\item{upper_change}{When not c(Inf), cells will be trimmed based on upper_thresholds}

\item{lower_thresholds}{Default of NULL only allowed when change is Inf}

\item{upper_thresholds}{Default of NULL only allowed when change is Inf}

\item{cell_size}{Cell size used in make-vector-data function}

\item{dist_intercept}{Distance at which model estimates will be calculated.
Defaults to nearest neighbouring cells}

\item{max_dist}{Value at which to truncate distances (defaults to no truncation)}

\item{min_dist}{Defaults to half cell size.}
}
\description{
Trim VOCC output to only include vectors for cells with specific end conditions
}
