% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc-gradient.R
\name{vocc_gradient_calc}
\alias{vocc_gradient_calc}
\title{Velocity of climate change}
\usage{
vocc_gradient_calc(
  data,
  layer,
  scale_fac = 1,
  time_step = "year",
  delta_t_step = 2,
  indices = NULL,
  divisor = 10,
  latlon = FALSE,
  log_space = FALSE,
  use_VoCC = F,
  quantile_cutoff = 0.05
)
}
\arguments{
\item{data}{Gridded data.}

\item{layer}{Variable to retain. Defaults to "est", for raw sdmTMB predictions.}

\item{scale_fac}{Controls how the original 2-km projection is aggregated
(e.g. a value of 5 means that the raster would be reprojected to 10km grid).}

\item{time_step}{Time variable.}

\item{delta_t_step}{Time between values of time variable. 
Required if raster layers not labeled with "X" plus an integer representing time.}

\item{indices}{If null, gradient will be based on mean across all times. 
If integer vector of same length as number of unique time steps, 
gradient will be based on final time steps given largest index value.}

\item{divisor}{Default of 10 gives trends and velocities per decade.}

\item{latlon}{Default is TRUE. False for UTMs or other true distance grid.}

\item{log_space}{Logical for layer being in log space. 
If TRUE will save 'cv' in log space and true 'sd'.}

\item{quantile_cutoff}{Used for plotting to set min and max angles of vectors.
Defaults to 0.05.}
}
\description{
Velocity of climate change
}
